@@ -27,7 +27,8 @@
     "url": "https://github.com/gribnoysup/react-yandex-maps.git"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
+    "react": "^0.14.0 || ^15.5.0",
+    "prop-types": " ^15.5.0"
   },
   "devDependencies": {
     "babel-cli": "^6.23.0",
@@ -60,7 +61,8 @@
     "prettier": "^0.21.0",
     "pretty-bytes": "^4.0.2",
     "raw-loader": "^0.5.1",
-    "react": "^15.4.2",
+    "react": "^15.5.0",
+    "prop-types": " ^15.5.0",
     "react-ace": "^4.1.6",
     "react-dom": "^15.4.2",
     "react-router": "^4.0.0-beta.6",
